Option to apply temp basal rate depending on glucose value, for a predefined period.
This is an option to handle the unpredictable effects of meals (in other words : fats).
It will apply a temporary basal rate, dependent on current glucose value.

Below a table with rates being used, normalized to 1 unit/hour when values is above 250 and trend is Up

These rates can be multiplied with a factor 1 to 3.8 via the xDrip4iOS UI

Also meal duration can be set, 1, 2, 3, 4, 5 or 6 hours.
There's two columns : first column is used first half of the duration, second column is used second half of the duration.

If the auto basal algorithm recommends a temp basal, then there will be no micro-bolus recommended

Changes in Loop and xdrip-client-swift
https://github.com/JohanDegraeve/Loop-LoopKit/commits/mydev2
https://github.com/JohanDegraeve/xdrip-client-swift-1/commits/mydev

Range	Trend			Max first half	Max Second half
120-160	UpUpUp, UpUp		0,3145728	0,2883584
	Up			0,262144	0,2359296
	flat			0,2097152	0,1835008
	Down			0,1048576	0,0786432
	DownDown, DownDownDown	0		0
160-180	UpUpUp, UpUp		0,49152		0,45056
	Up			0,4096		0,36864
	flat			0,32768		0,28672
	Down			0,16384		0,12288
	DownDown, DownDownDown	0		0
180-200	UpUpUp, UpUp		0,6144		0,5632
	Up			0,512		0,4608
	flat			0,4096		0,3584
	Down			0,2048		0,1536
	DownDown, DownDownDown	0		0
200-220	UpUpUp, UpUp		0,768		0,704
	Up			0,64		0,576
	flat			0,512		0,448
	Down			0,256		0,192
	DownDown, DownDownDown	0		0
220-250	UpUpUp, UpUp		0,96		0,88
	Up			0,8		0,72
	flat			0,64		0,56
	Down			0,32		0,24
	DownDown, DownDownDown	0		0
> 250	UpUpUp, UpUp		1,2		1,1
	Up			1,0		0,9
	flat			0,8		0,7
	Down			0,4		0,3
	DownDown, DownDownDown	0		0